requirements.txt
charset-normalizer==2.0.6

click==8.0.1

flask==2.0.1

idna==3.2

itsdangerous==2.0.1

jinja2==3.0.1

jmespath==0.10.0

markupsafe==2.0.1

pyjwt==2.1.0

python-dateutil==2.8.2

mysql-connector==2.2.9

mysql-connector-python~=8.0.16

requests==2.26.0

s3transfer==0.5.0

gunicorn==20.0.4

werkzeug==2.0.1

boto3==1.18.53

botocore==1.21.53

twilio==7.0.0

six==1.16.0

urllib3==1.26.7

setuptools~=56.0.0

pip~=21.1.3

wsl~=0.12

pytz~=2021.3

lib~=3.0.0

config~=0.5.0.post0

wheel~=0.37.0

python-dotenv~=0.19.1

protobuf~=3.17.3

docutils~=0.17.1

ratelimit~=2.2.1
